requirements.txt
lxml==4.5.0
numpy==1.18.2
opc-diag==1.0.0
pandas==1.0.3
pyspellchecker==0.5.4
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
six==1.14.0
xlrd==1.2.0
XlsxWriter==1.2.8
